Landing
Implementing different states for ship. For now: Flying (1) (The normal state until now) and Landing (2).
Probably I will add two more: Rotating(3) So the ship always faces in the right direction after landing, and Takingoff(4) When the ship is ready to takeoff and go flying around space again. This is quite experimental. I hope this works.